+++ CHANGELOG.md
@@ -26,6 +26,8 @@ CHANGELOG
   insufficient space on the right side.
 - Fix #107: Remove the dropdown after timeout to fix "hover" behavior.
 - Fix #136: Update original `<select>` element on "change" instead of "selectivity-selected".
+- Fix: When a Selectivity instance is clicked but its dropdown should not open, at least it should
+       be focused.
 
 ## 2.1.0